==> 1_Separate train_val_test (No Resizing).ipynb <==
import numpy as np
import pandas as pd
from scipy import misc

import os
import shutil
from tqdm import tqdm 

from PIL import Image

# the folder for images
dir_train = 'E:/Jupyter NB/10 class Classification/Images'

# the directory of separeated images
image_dir = 'E:/Jupyter NB/10 class Classification/Train_val_test'

# This cell is for testing

'''
for root, dirs, files in os.walk(dir_train):
    print('---root---')
    print(root)
    print('---dirs---')
    print(dirs)
    print('---files---')
    print(files)
    print('---------------------')



subdirs = list(os.walk(dir_train))[1:]

for dir_path, _, files in tqdm(subdirs):
    
    dir_name = dir_path.split('\\')[-1]
    print(dir_name)
'''

subdirs = list(os.walk(dir_train))[1:]

# collect train metadata
train_metadata = []

for dir_path, _, files in tqdm(subdirs):
    
    dir_name = dir_path.split('\\')[-1]
    
    for file_name in files:
        if not file_name.startswith('.'):
            # read image
            temp = misc.imread(os.path.join(dir_path, file_name)) 
            # collect image metadata
            image_metadata = []
            image_metadata.extend([dir_name, file_name])
            image_metadata.extend( 
                list(temp.shape) if len(temp.shape) == 3 
                else [temp.shape[0], temp.shape[1], 1]
            )
            image_metadata.extend([temp.nbytes, temp.dtype])
            # append image metadata to list
            train_metadata.append(image_metadata)

#print(train_metadata)

M = pd.DataFrame(train_metadata)
M.columns = ['directory', 'img_name', 'height', 'width', 'channels', 'byte_size', 'bit_depth']

M['category_name'] = M.directory.apply(lambda x: x.split('.')[-1].lower())
M['img_extension'] = M.img_name.apply(lambda x: x.split('.')[-1])
M['category_number'] = M.directory.apply(lambda x: int(x.split('.')[0]))

# number of grayscale images
(M.channels != 3).sum()

# number of categories
M.category_name.nunique()

# class number -> class name
decode = {n: i for i, n in M.groupby('category_name').category_number.first().iteritems()}
decode

np.save('decode.npy', decode)

# 20 images per class
V = M.groupby('category_name', group_keys=False).apply(lambda x: x.sample(n=20, replace=False))
V.sort_index(inplace=True)
M.drop(V.index, axis=0, inplace=True)

'''
# 20 images per class for test
Test = M.groupby('category_name', group_keys=False).apply(lambda x: x.sample(n=20, replace=False))
Test.sort_index(inplace=True)
M.drop(Test.index, axis=0, inplace=True)
'''

# train data
len(M)

# validation data
len(V)

#Test data
#len(Test)

M.to_csv('train_metadata.csv', index=False)
V.to_csv('val_metadata.csv', index=False)
#Test.to_csv('test_data.csv', index=False)

t = pd.read_csv('E:/Jupyter NB/10 class Classification/train_metadata.csv')
v = pd.read_csv('E:/Jupyter NB/10 class Classification/val_metadata.csv')
#test = pd.read_csv('E:/Jupyter NB/10 class Classification/test_data.csv')

os.mkdir(image_dir + '/train_no_resizing')
for i in range(1, 20 + 1):
    os.mkdir(image_dir + '/train_no_resizing/' + str(i))

os.mkdir(image_dir + '/val_no_resizing')
for i in range(1, 20 + 1):
    os.mkdir(image_dir + '/val_no_resizing/' + str(i))

'''
os.mkdir(image_dir + '/test_no_resizing')
for i in range(1, 20 + 1):
    os.mkdir(image_dir + '/test_no_resizing/' + str(i))
'''

# Train images

train_size = len(t)
train_size

# RGB images
for i, row in tqdm(t.loc[t.channels == 3].iterrows()):
    # get image
    file_path = os.path.join(dir_train, row.directory, row.img_name)
    image = Image.open(file_path)
    
    # save
    save_path = os.path.join(image_dir, 'train_no_resizing', str(row.category_number), row.img_name)
    image.save(save_path, 'jpeg')

# grayscale images
for i, row in tqdm(t.loc[t.channels == 1].iterrows()):
    # get image
    file_path = os.path.join(dir_train, row.directory, row.img_name)
    image = Image.open(file_path)
    
    # convert to RGB
    array = np.asarray(image, dtype='uint8')
    array = np.stack([array, array, array], axis=2)
    image = Image.fromarray(array)
    
    # save
    save_path = os.path.join(image_dir, 'train_no_resizing', str(row.category_number), row.img_name)
    image.save(save_path, 'jpeg')

# Validation images

val_size = len(v)
val_size

# RGB images
for i, row in tqdm(v.loc[v.channels == 3].iterrows()):
    # get image
    file_path = os.path.join(dir_train, row.directory, row.img_name)
    image = Image.open(file_path)
    
    # save
    save_path = os.path.join(image_dir, 'val_no_resizing', str(row.category_number), row.img_name)
    image.save(save_path, 'jpeg')

# grayscale images
for i, row in tqdm(v.loc[v.channels == 1].iterrows()):
    # get image
    file_path = os.path.join(dir_train, row.directory, row.img_name)
    image = Image.open(file_path)
    
    # convert to RGB
    array = np.asarray(image, dtype='uint8')
    array = np.stack([array, array, array], axis=2)
    image = Image.fromarray(array)
    
    # save
    save_path = os.path.join(image_dir, 'val_no_resizing', str(row.category_number), row.img_name)
    image.save(save_path, 'jpeg')

# Test images

'''
test_size = len(test)
test_size
'''

'''
# RGB images
for i, row in tqdm(test.loc[test.channels == 3].iterrows()):
    # get image
    file_path = os.path.join(dir_train, row.directory, row.img_name)
    image = Image.open(file_path)
    
    # save
    save_path = os.path.join(image_dir, 'test_no_resizing', str(row.category_number), row.img_name)
    image.save(save_path, 'jpeg')
'''

'''
# grayscale images
for i, row in tqdm(test.loc[test.channels == 1].iterrows()):
    # get image
    file_path = os.path.join(dir_train, row.directory, row.img_name)
    image = Image.open(file_path)
    
    # convert to RGB
    array = np.asarray(image, dtype='uint8')
    array = np.stack([array, array, array], axis=2)
    image = Image.fromarray(array)
    
    # save
    save_path = os.path.join(image_dir, 'test_no_resizing', str(row.category_number), row.img_name)
    image.save(save_path, 'jpeg')
'''
